Create NFCUtils
