+++ views/Send.tsx
@@ -31,6 +31,7 @@ import UTXOPicker from './../components/UTXOPicker';
 import FeeTable from './../components/FeeTable';
 
 import RESTUtils from './../utils/RESTUtils';
+import NFCUtils from './../utils/NFCUtils';
 import { localeString } from './../utils/LocaleUtils';
 import { themeColor } from './../utils/ThemeUtils';
 
@@ -122,37 +123,6 @@ export default class Send extends React.Component<SendProps, SendState> {
         await this.initNfc();
     }
 
-    // TODO consider making an NFC util
-    utf8ArrayToStr = (data: any) => {
-        const extraByteMap = [1, 1, 1, 1, 2, 2, 3, 0];
-        const count = data.length;
-        let str = '';
-
-        for (let index = 0; index < count; ) {
-            const ch = data[index++];
-            if (ch & 0x80) {
-                const extra = extraByteMap[(ch >> 3) & 0x07];
-                if (!(ch & 0x40) || !extra || index + extra > count) {
-                    return null;
-                }
-
-                ch = ch & (0x3f >> extra);
-                for (; extra > 0; extra -= 1) {
-                    const chx = data[index++];
-                    if ((chx & 0xc0) !== 0x80) {
-                        return null;
-                    }
-
-                    ch = (ch << 6) | (chx & 0x3f);
-                }
-            }
-
-            str += String.fromCharCode(ch);
-        }
-
-        return str;
-    };
-
     initNfc = async () => {
         await NfcManager.start();
 
@@ -167,8 +137,7 @@ export default class Send extends React.Component<SendProps, SendState> {
             NfcManager.setEventListener(NfcEvents.DiscoverTag, (tag: any) => {
                 tagFound = tag;
                 const bytes = new Uint8Array(tagFound.ndefMessage[0].payload);
-                // TODO handle this more elegantly
-                const str = this.utf8ArrayToStr(bytes).slice(3);
+                const str = NFCUtils.nfcUtf8ArrayToStr(bytes);
                 resolve(this.validateAddress(str));
                 NfcManager.unregisterTagEvent().catch(() => 0);
             });
